refactor: error 관련 패키지명 수정
