+++ api/src/main/java/com/tago/api/auth/jwt/filter/JwtFilter.java
@@ -3,8 +3,8 @@ package com.tago.api.auth.jwt.filter;
 import com.tago.api.auth.jwt.JwtProvider;
 import com.tago.api.auth.jwt.exception.AuthenticateException;
 import com.tago.api.auth.jwt.handler.JwtAuthenticationEntryPoint;
-import com.tago.domain.common.exception.BaseBusinessException;
-import com.tago.domain.common.exception.ErrorCode;
+import com.tago.domain.common.error.BaseBusinessException;
+import com.tago.domain.common.error.ErrorCode;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
