+++ domain/src/main/java/com/tago/domain/common/error/BaseBusinessException.java
@@ -1,4 +1,4 @@
-package com.tago.domain.common.exception;
+package com.tago.domain.common.error;
 
 import lombok.Getter;
