+++ domain/src/main/java/com/tago/domain/common/error/ErrorCode.java
@@ -1,4 +1,4 @@
-package com.tago.domain.common.exception;
+package com.tago.domain.common.error;
 
 import lombok.Getter;
